Added Room Database, ShoppingListsFragment and navigation graph
Room Database:
Entities - ShoppingList and ShoppingListItem, with one-to-many relation
(ShoppingListWithItems).
Database and repository are bound to global application state
(in ShoppingListApp) so that they are accessible in all fragments and
activities.

Added ShoppingListsFragment together with ShoppingListsAdapter for the
recyclerview and ShoppingListsViewModel for accessing the data.

Added the navigation graph and updated activity_main.xml to include it.